Rename endian.hpp to endian_read.hpp
Following #7620, which split the write functions into a separate header.
