+++ Source/codec.cpp
@@ -5,7 +5,7 @@
 
 #include "appfat.h"
 #include "sha.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/log.hpp"
 
 namespace devilution {
