+++ Source/diablo.h
@@ -12,7 +12,7 @@
 #endif
 #include "levels/gendung.h"
 #include "utils/attributes.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 
 namespace devilution {
