+++ Source/dvlnet/frame_queue.cpp
@@ -5,7 +5,7 @@
 #include "appfat.h"
 #include "dvlnet/packet.h"
 #include "utils/attributes.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/endian_write.hpp"
 
 namespace devilution {
