+++ Source/engine/clx_sprite.hpp
@@ -30,7 +30,7 @@
 #include <memory>
 
 #include "appfat.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/intrusive_optional.hpp"
 
 namespace devilution {
