+++ Source/loadsave.cpp
@@ -37,7 +37,7 @@
 #include "qol/stash.h"
 #include "stores.h"
 #include "utils/algorithm/container.hpp"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/is_of.hpp"
 #include "utils/language.h"
 #include "utils/status_macros.hpp"
