+++ Source/mpq/mpq_writer.cpp
@@ -11,7 +11,7 @@
 #include "appfat.h"
 #include "encrypt.h"
 #include "engine.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/file_util.h"
 #include "utils/language.h"
 #include "utils/log.hpp"
