+++ Source/pack.cpp
@@ -14,7 +14,7 @@
 #include "playerdat.hpp"
 #include "plrmsg.h"
 #include "stores.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/is_of.hpp"
 #include "utils/log.hpp"
 #include "utils/utf8.hpp"
