+++ Source/pfile.cpp
@@ -23,7 +23,7 @@
 #include "pack.h"
 #include "playerdat.hpp"
 #include "qol/stash.h"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/file_util.h"
 #include "utils/language.h"
 #include "utils/parse_int.hpp"
