+++ Source/utils/cel_to_clx.cpp
@@ -12,7 +12,7 @@
 
 #include "appfat.h"
 #include "utils/clx_encode.hpp"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/endian_write.hpp"
 
 namespace devilution {
