+++ Source/utils/cl2_to_clx.cpp
@@ -7,7 +7,7 @@
 
 #include "utils/clx_decode.hpp"
 #include "utils/clx_encode.hpp"
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/endian_write.hpp"
 
 namespace devilution {
