+++ Source/utils/endian_stream.hpp
@@ -6,7 +6,7 @@
 #include <cstdio>
 #include <cstring>
 
-#include "utils/endian.hpp"
+#include "utils/endian_read.hpp"
 #include "utils/endian_write.hpp"
 #include "utils/log.hpp"
